Step 3: Add very basic CRUD API for Todos (MongoDB usage, Interceptors error handling, Controller validation) (#3)
* chore: install mongodb and mongoose

* feat: implement the same crud api via mongoose approach instead of file-system approach

* feat: moving secrets into dotenv file

* chore: rename model and related code

* refactor: routing and database constants

* fix: add await to all api methods, remove returned todo from delete method

* fix: add _id to the todo schema

* refactor: use nestjs-config instead of dotenv

* chore: rename mongodb uri env variable

* refactor: use nestjs-mongoose schema instead of mongoose-schema

* feat: handle error via the global interceptor

* fix: custom error classes to handle errors in a more correct way

* chore: update ts configuration, add env example file

* fix: expand custom error classes with all args, add env variable check in app module

* fix: update todo schema

* feat: add controller validation (manually written)

* refactor: use multiple if statements instead of switch, move logic inside throw-error callback

* fix: encapsulate all todos array by converting it from model into object

+++ src/todos/todos.controller.ts
@@ -3,43 +3,86 @@ import {
   Controller,
   Delete,
   Get,
+  HttpCode,
   Param,
   Patch,
   Post,
 } from '@nestjs/common'
+import { Types } from 'mongoose'
 
+import { TODOS_ROUTE } from 'src/common/constants/routing.constants'
+import { ValidationError } from 'src/common/errors/errors'
 import { TodosService } from './todos.service'
-import type { TodoItem } from 'types/todos'
+import type { Todo } from './schemas/todos.schema'
 
-@Controller('todos')
+@Controller(TODOS_ROUTE)
 export class TodosController {
   constructor(private readonly todosService: TodosService) {}
 
   @Post()
-  create(@Body() todo: TodoItem): TodoItem {
-    return this.todosService.create(todo)
+  async create(@Body() todo: any): Promise<Todo> {
+    if (
+      typeof todo !== 'object' ||
+      typeof todo.title !== 'string' ||
+      !todo.title.trim()
+    ) {
+      throw new ValidationError(
+        'Invalid todo. Title is required and should be a non-empty string.',
+      )
+    }
+
+    return await this.todosService.create(todo)
   }
 
   @Get()
-  getAll(): TodoItem[] {
-    return this.todosService.getAll()
+  async getAll(): Promise<Todo[]> {
+    return await this.todosService.getAll()
   }
 
   @Get(':id')
-  getById(@Param('id') id: string): TodoItem {
-    return this.todosService.getById(id)
+  async getById(@Param('id') id: any): Promise<Todo> {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new ValidationError(
+        'Invalid id. ID should be a valid MongoDB ObjectId.',
+      )
+    }
+
+    return await this.todosService.getById(id)
   }
 
   @Patch(':id')
-  update(
-    @Param('id') id: string,
-    @Body() newParams: Partial<TodoItem>,
-  ): TodoItem {
-    return this.todosService.update(id, newParams)
+  async update(@Param('id') id: any, @Body() newParams: any): Promise<Todo> {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new ValidationError(
+        'Invalid id. ID should be a valid MongoDB ObjectId.',
+      )
+    }
+    if (
+      newParams.title &&
+      (typeof newParams.title !== 'string' || !newParams.title.trim())
+    ) {
+      throw new ValidationError(
+        'Invalid title. Title should be a non-empty string.',
+      )
+    }
+    if (newParams.description && typeof newParams.description !== 'string') {
+      throw new ValidationError(
+        'Invalid description. Description should be a string.',
+      )
+    }
+
+    return await this.todosService.update(id, newParams)
   }
 
   @Delete(':id')
-  delete(@Param('id') id: string): TodoItem {
-    return this.todosService.delete(id)
+  @HttpCode(204)
+  async delete(@Param('id') id: any): Promise<void> {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new ValidationError(
+        'Invalid id. ID should be a valid MongoDB ObjectId.',
+      )
+    }
+
+    await this.todosService.delete(id)
   }
 }
